Account for melting red ice walls being permanent
This puts melting the red ice walls in Ice Cavern using blue fire arrows and then coming back as child to collect the skulltulas and red rupees with boomerang into logic. I checked all uses of the `Blue_Fire` logic helper and this seems to be the only case where this is relevant.