Track state items in a flat list.
Since we have a fixed list of progress items (including event items),
it's a lot more time-efficient to encode them as an enum-like integer
type and use them as indexes into a fixed-size list, than to use
Counter--especially for "cache-miss" lookups of items we don't have in
the dict.

It doesn't really make a difference whether we use an int constant or a
top-level Name, so we just give all the items a unique id and a variable
name (that matches what we write in the logic, ha!) and then put those
in the global scope for the access rule.

Also:
- Improved all_of and any_of since the functional code isn't as fast
  on lists as just looping.
- Fixed visiting locations in the wrong world copy during playthrough.
- Added Epona to the items list to solve an issue with fast epona.
  This will also add Epona to playthroughs.